requirements.txt
agate==1.6.3
alembic==1.6.5
aniso8601==7.0.0
anyio==3.6.1
asn1crypto==1.5.1
attrs==22.1.0
Babel==2.10.3
beautifulsoup4==4.11.1
bleach==5.0.1
build==0.8.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
coloredlogs==14.0
croniter==1.3.5
cryptography==36.0.2
dagit==1.0.3
dagster==1.0.3
dagster-dbt==0.16.3
dagster-graphql==1.0.3
dagster-pandas==0.16.3
dagster-snowflake==0.16.3
dbt-core==1.2.0
dbt-extractor==0.4.1
dbt-snowflake==1.2.0
defusedxml==0.7.1
docstring-parser==0.14.1
entrypoints==0.4
fastjsonschema==2.16.1
future==0.18.2
gql==2.0.0
graphene==2.1.9
graphql-core==2.3.2
graphql-relay==2.0.1
grpcio==1.47.0
grpcio-health-checking==1.43.0
h11==0.13.0
hologram==0.0.15
httptools==0.4.0
humanfriendly==10.0
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.9.0
iniconfig==1.1.1
isodate==0.6.1
Jinja2==3.1.2
jsonschema==3.2.0
jupyter-client==7.3.4
jupyter-core==4.11.1
jupyterlab-pygments==0.2.2
keyring==23.8.2
leather==0.3.4
Logbook==1.5.3
lxml==4.9.1
Mako==1.2.1
MarkupSafe==2.0.1
mashumaro==2.9
minimal-snowplow-tracker==0.0.2
mistune==0.8.4
msgpack==1.0.4
mypy-extensions==0.4.3
nbclient==0.6.6
nbconvert==6.5.3
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==2.8.5
numpy==1.23.1
oscrypto==1.3.0
packaging==21.3
pandas==1.4.3
pandocfilters==1.5.0
parsedatetime==2.4
pathspec==0.9.0
pendulum==2.1.2
pep517==0.13.0
pkgutil_resolve_name==1.3.10
platformdirs==2.5.2
pluggy==1.0.0
promise==2.3
protobuf==3.20.1
py==1.11.0
pycparser==2.21
pycryptodomex==3.15.0
Pygments==2.12.0
PyJWT==2.4.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyrsistent==0.18.1
pytest==7.1.2
python-dateutil==2.8.2
python-dotenv==0.20.0
python-editor==1.0.4
python-slugify==6.1.2
pytimeparse==1.1.8
pytz==2022.2.1
pytzdata==2020.1
PyYAML==6.0
pyzmq==23.2.1
requests==2.28.1
Rx==1.6.1
six==1.16.0
sniffio==1.2.0
snowflake-connector-python==2.7.11
soupsieve==2.3.2.post1
SQLAlchemy==1.4.40
sqlparse==0.4.2
starlette==0.20.4
tabulate==0.8.10
text-unidecode==1.3
tinycss2==1.1.1
tomli==2.0.1
toposort==1.7
tornado==6.2
tqdm==4.64.0
traitlets==5.3.0
typing-compat==0.1.0
typing_extensions==4.3.0
urllib3==1.26.11
uvicorn==0.18.2
uvloop==0.16.0
watchdog==2.1.9
watchfiles==0.16.1
webencodings==0.5.1
websockets==10.3
Werkzeug==2.1.2
zipp==3.8.1
